requirements.txt
python-binance==1.0.19
openai==1.12.0
pandas==2.2.0
numpy==1.26.4
schedule==1.2.1
pytz==2024.1
requests==2.31.0
urllib3==2.2.1
python-dotenv==1.0.0